Подписал тесты, разрешил им использовать Internal метотд из основной сборки
