+++ src/ConfigSettings/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
@@ -14,3 +15,5 @@ using System.Runtime.InteropServices;
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("2411fc38-901f-4e58-ab3f-0ff0f5236766")]
+
+[assembly: InternalsVisibleTo("ConfigSettingsTests, PublicKey=002400000480000094000000060200000024000052534131000400000100010001c52a3121846a9b90a3366b68fd0fc4dbc160b5455c169704928e1b9c51eec670945d5dbfcc6dd28328036cfa38e1af60b51abd81a6376cf6ecadb7f65e098a17105adbc48b2157eb08e3e077ce15ea30e3fa7c1c9349347f4c6f572346f9bfb7cc8d71e6d3dd3261ecb945781f584ecd689a650f05bdbef9b80a420eb2a4d5")]
