+++ src/Tests/ConfigSettingsGetterTests.cs
@@ -52,66 +52,6 @@ namespace ConfigSettings.Tests
     }
 
     [Test]
-    public void WhenChangeVariableThenBlockShouldNotChanged()
-    {
-      var configSettingsPath = this.CreateSettings(@"   <var name=""GIT_ROOT_DIRECTORY"" value=""d:\ee"" />
-  <block name=""REPOSITORIES"">
-    <repository folderName=""base"" solutionType=""Base"" url="""" />
-    <repository folderName=""work"" solutionType=""Work"" url="""" />
-  </block>");
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.Set("GIT_ROOT_DIRECTORY", "d:\\ee2");
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <var name=""GIT_ROOT_DIRECTORY"" value=""d:\ee2"" />
-  <block name=""REPOSITORIES"">
-    <repository folderName=""base"" solutionType=""Base"" url="""" />
-    <repository folderName=""work"" solutionType=""Work"" url="""" />
-  </block>
-");
-    }
-
-    [Test]
-    public void WhenSetEmptyBlockContent()
-    {
-      var configSettingsPath = this.CreateSettings(@"  <block name=""REPOSITORIES"">
-  </block>");
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.SetBlock("TESTBLOCK", null, null);
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <block name=""REPOSITORIES""></block>
-  <block name=""TESTBLOCK""></block>
-");
-    }
-
-    [Test]
-    public void WhenSetEmptyBlockEnabled()
-    {
-      var configSettingsPath = this.CreateSettings(@"  
-  <block name=""ORIGIN_TRUE_BLOCK"" enabled=""True""/>
-  <block name=""ORIGIN_FALSE_BLOCK"" enabled=""false""/>
-  <block name=""ORIGIN_NULL_BLOCK""/>
-");
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.SetBlock("TEST_TRUE_BLOCK", true, null);
-      getter.SetBlock("TEST_FALSE_BLOCK", false, null);
-      getter.SetBlock("TEST_NULL_BLOCK", null, null);
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <block name=""ORIGIN_TRUE_BLOCK"" enabled=""true""></block>
-  <block name=""ORIGIN_FALSE_BLOCK"" enabled=""false""></block>
-  <block name=""ORIGIN_NULL_BLOCK""></block>
-  <block name=""TEST_TRUE_BLOCK"" enabled=""true""></block>
-  <block name=""TEST_FALSE_BLOCK"" enabled=""false""></block>
-  <block name=""TEST_NULL_BLOCK""></block>
-");
-    }
-
-    [Test]
     public void WhenGetEmptyXmlBlock()
     {
       var configSettingsPath = this.CreateSettings(@"  
@@ -161,83 +101,6 @@ namespace ConfigSettings.Tests
       thrownException.Should().BeOfType<ParseConfigException>();
       thrownException.InnerException.Should().BeOfType<XmlException>();
       ((ParseConfigException)thrownException).CorruptedFilePath.Should().Be(configSettingsPath);
-    }
-
-    [Test]
-    public void WhenSetBlockValueAndEnabled()
-    {
-      var configSettingsPath = this.CreateSettings(@"  
-  <block name=""ORIGIN_TRUE_BLOCK"" enabled=""false"">
-    <repository folderName=""base"" solutionType=""Base"" url="""" />
-    <repository folderName=""work"" solutionType=""Work"" url="""" />
-</block>
-");
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.SetBlock("TEST_TRUE_BLOCK", true, @"
-  <testRepository folderName=""base"" solutionType=""Base"" url="""" />
-  <testRepository folderName=""work"" solutionType=""Work"" url="""" />");
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <block name=""ORIGIN_TRUE_BLOCK"" enabled=""false"">
-    <repository folderName=""base"" solutionType=""Base"" url="""" />
-    <repository folderName=""work"" solutionType=""Work"" url="""" />
-  </block>
-  <block name=""TEST_TRUE_BLOCK"" enabled=""true"">
-    <testRepository folderName=""base"" solutionType=""Base"" url="""" />
-    <testRepository folderName=""work"" solutionType=""Work"" url="""" />
-  </block>
-");
-    }
-
-    [Test]
-    public void WhenSetRelativeImportThenPathShouldNotBeAbsolute()
-    {
-      var configSettingsPath = this.CreateSettings(@"  
-  <import from=""origin/import/from"" />
-");
-
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.SetImport("test/import/from");
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <import from=""origin/import/from"" />
-  <import from=""test/import/from"" />
-");
-    }
-
-    [Test]
-    public void WhenParseEmptyImportAndAddVariableThenImportBlockShouldNotBeSaved()
-    {
-      var configSettingsPath = this.CreateSettings(@"  
-  <import from="""" />
-");
-      var getter = CreateConfigSettingsGetter(configSettingsPath);
-      getter.Set("testName", "testValue");
-      getter.Save();
-      var content = TestTools.GetConfigSettings(configSettingsPath);
-      content.Should().Be(@"
-  <var name=""testName"" value=""testValue"" />
-");
-    }
-
-    [Test]
-    public void WhenSetValueWithNullConfigSettingsParserThenExceptionShouldBeRaisen()
-    {
-      var getter = CreateConfigSettingsGetter(null);
-      getter.Get<bool>("NEW_VALUE").Should().BeFalse();
-      getter.Set("NEW_VALUE", true);
-      getter.Get<bool>("NEW_VALUE").Should().BeTrue();
-    }
-
-    [Test]
-    public void WhenSaveNullConfigSettingsParserThenSaveShouldNotWork()
-    {
-      var getter = CreateConfigSettingsGetter(null);
-      getter.Set("NEW_VALUE", true);
-      var exception = ((MethodThatThrows)delegate { getter.Save(); }).GetException();
-      exception.Should().BeOfType<InvalidOperationException>();
     }
 
     [Test]
