+++ src/Tests/ConfigSettingsTests.csproj
@@ -10,6 +10,8 @@
     <StartupObject />
     <PlatformTarget>AnyCPU</PlatformTarget>
     <Product>Test Product Name</Product>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>ConfigSettingsTest.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Properties\AssemblyInfo.cs" />
